requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai==1.0.0
pillow==10.2.0
